requirements.txt
djangorestframework==3.16.0
drf-yasg==1.21.10
pillow==10.4.0
django-parler